# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice for scraping anime data",
        "title": "Anilibrary Scraper",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Boost Software License, Version 1.0",
            "url": "https://github.com/VampireAotD/anilibrary-scraper/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/anime/parse": {
            "post": {
                "description": "Scrape anime data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anime"
                ],
                "summary": "Scrape anime data",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Url to scrape from",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ScrapeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ScrapeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.ScrapeRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "response.Entry": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ScrapeResponse": {
            "type": "object",
            "properties": {
                "episodes": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Entry"
                    }
                },
                "image": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "synonyms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Entry"
                    }
                },
                "title": {
                    "type": "string"
                },
                "voiceActing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Entry"
                    }
                }
            }
        }
    }
}